[SYCL][RTC] Cache frontend invocation (#16823)
Adds `sycl_jit`-RTC-specific persistent caching to the runtime. The
basic idea is to cache only the LLVM module resulting from the device
compilation in bitcode format. Device linking and post-link would be run
always (even for a cache hit), as invoking the frontend is the most
expensive step in the pipeline right now.

The cache key is the concatenation of:
- the Base64*-encoding of a BLAKE3 hash of the preprocessed source
string (i.e. containing all headers included as virtual files per the
`kernel_compiler` extension as well as from the local file system), and
- the Base64-encoding of a BLAKE3 hash of the user-supplied build
options.


*) Replacing `/` by `-` to make the string filesystem-friendly.

---------

Signed-off-by: Julian Oppermann <julian.oppermann@codeplay.com>
Co-authored-by: Lukas Sommer <lukas.sommer@codeplay.com>