[SYCL][CUDA] Fix flaky interop task test
In host tasks, `Q.wait()` doesn't wait on interop asynchronous work, it
simply waits on the lambba being executed.

So for correct execution the CUDA stream must be waited on inside of the
lambda.

This made the test flaky because in some cases the copies had been
completed before being checked, and in some cases they hadn't, this
extra wait fixes that.