[libspirv] Fix target dependency on SPIR-V conversion builtins
Upstream commit 15c2d1b fixed dependencies on the generated conversion
builtins by automatically identifying targets that produce a generated
files. It does so by taking the file name and finding a target with the
same name but a prefix 'generate-'.

When this commit was pulled down into DPC++, this was missed and so the
target dependencies were incorrect on the (downstream) SPIR-V conversion
builtins.