[SYCL][LIBCLC] Use half builtins for generic math functions
`__builtin_<name>()` without a suffix is equivalent to the C math
function `<name>()`, these functions take and return `double` values.

Instead of using these use the `f16` suffixed builtins which take and
return `half` values.

On CPU targets these will be usually lowered to
the single precision library calls (with promotion/truncation). On other
targets (for example GPUs) the half builtins might be lowered directly
to hardware instructions.

Some of the builtins are not available for half precision, in these cases
use the `f` suffixed builtins (which take and return `float`s).