[UR] Add reference counting and ext. fn cache to OpenCL adapters
Rather than using atexit handles to drop the adapter (which causes
ordering issues with SYCL atexit handlers),
`urAdapterRetain/Release` now update a reference counter. When the
last adapter handle is Released, the adapter is dropped.

Since it is now simple to do so, the function cache has been moved
to be a member of the adapter itself rather than being a separate
global.