[libspirv] Restore overloads of generic address space
This adds some missing generic address space overloads of CLC functions:
frexp, and whichever functions make use of the
unary_def_with_(int_)?ptr.inc files.

Note that this commit doesn't map the equivalent SPIR-V functions to
these CLC implementations. This is because the generic SPIR-V
implementation of FP16 types currently use the clang builtins as opposed
to the software implementations. As discussed in #17163, this is not the
proper behaviour for generic implementations, as not all targets support
the builtins. In future work we will need to either provide
target-specific overrides for the builtin implementations, or provide an
easy mechanism by which the CLC implementations can select between
builtins or software implementations automatically.