[SYCL][E2E] Make 8 free checks accounting for page query optional
When the L0 adapter internal pooling is turned off
eg. with SYCL_PI_LEVEL_ZERO_DISABLE_USM_ALLOCATOR=1, the adapter no
longer creates an UMF pool that queries for page size during its
creation. The allocations made for the query were accounted for in
the L0 buffer ownership test, and it's no longer the case. It still
affects the v2 L0 adapter.