[SYCL][CUDA] Fix cupti library dynamic loading
This patch fixes loading the cupti library on Linux to rely on the
dynamic linker to figure out where the library is.

It also disable cupti tracing on Windows, as far as I know this is only
used on Linux at the moment since the tracing tools don't support
Windows.