[SYCL][Graph][L0] Enable Copy engine support with in-order command lists (#16830)
Currently, copy engine is only available to use for certain commands in
the out-of-order command-buffer path. This PR adds required
synchronization between the 2 command lists, when both are in-order.
Hence, it enables using the copy engine with in-order path.

For the immediateAppend path, this only works for now if the environment
variable `UR_L0_USE_DRIVER_COUNTER_BASED_EVENTS=0` is set.