[SYCL][HIP] Disable legacy images on HIP
Legacy image support for HIP was initially hidden behind an environment
variable as it's not really supported, bindless images should be used
instead.

But after looking a bit closer at the existing support, it's pretty much
entirely broken, almost none of our existing legacy images tests even
compile for AMD.

So this patch fully disables the legacy images for HIP and cleans up all
the related `UNSUPPORTED`, in a lot of cases `target-amd` had to be kept
since the `build-only` mode doesn't know about the device aspects.

It also cleans up existing `UNSUPPORTED` tags for `cuda` in the same
tests.