[SYCL][UR][CUDA] Use FindCUDAToolkit CMake module instead of FindCUDA
find_package(CUDA) is deprecated since CMake 3.10 and the functionality
we need is provided by find_package(CUDAToolkit) since CMake 3.17.

Thanks to SYCL configuration now requiring CMake >3.20 (#13664), we can
rely on find_package(CUDAToolkit) to work in all setups.

Remove the deprecated calls and replace them with the recommended one.
Clean up all extra CMake code dealing with finding CUPTI as that is
also no longer needed (partially thanks to #17272). Replace all
variables from the old module with corresponding ones from the new one.

This solves multiple issues with finding libraries, notably including
the failure to find libcuda.so automatically on systems where the CUDA
driver is not installed and only the toolkit is available. This is a
reasonable use case for building DPC++ on a build machine without a
GPU and distributing for use on GPU machines.