[llvm-foreach] Avoid running llvm-foreach when waiting for child processes (#17260)
This resolves https://github.com/intel/llvm/issues/15177

Currently, we have a 'while' loop that keeps checking for child
processes to complete. This causes llvm-foreach to run and consume
resources when its child processes are active.
Instead, we can use blocking waits to wait for child processes. During
blocking waits. parent process is idle and does not consume resources.

I tested this on a fairly large program compilation (with AOT) and it
seems to work as expected.

No new tests are needed.
Thanks

---------

Signed-off-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>