[SYCL] Emit fpbuiltin version of function only for function with FP arguments. (#17253)
When the arguments of a function are of half types, such as
`sqrt(half_type v)` they are expanded to:
`load ptr addrspace(4), ptr addrspace(4) %a.addr.ascast, align 8, !tbaa
!5`
This will generate a crash in the BE due to type mismatch. The expected
type of the function is `float`.
This patch fixes the issue by restricting the emission of `fpbuiltin`
only to functions that take floating point values arguments.