[SYCL][E2E] minor test change to ease release (#17325)
The debug story for sycl-jit is not determined yet, so supporting the
`-g` flag for the kernel_compiler is not a strict requirement.
Meanwhile, on the SPIR-V side we are changing some opcodes to be
extensions. This upshot of all that is that between the open source and
closed source side, drivers and code gen, the -g flag, *when passed to
the kernel_compiler*, is leading to errors. This will resolve itself
once both the open source and closed source sides of the compiler, and
the matching drivers, have all made the switch to the same SPIRV
symbols. In the meantime, these tests are being adjusted so they pass in
all contexts.