[SYCL] Update spec constant handling for struct hierarchies (#17204)
After #16976, CTS failed to compile due to an assertion being raised in
`sycl-post-link`, more specifically while processing a `vec` spec
constants at
https://github.com/intel/llvm/blob/35fba198274c4a9f00c3f9de21cbd564242850c3/llvm/lib/SYCLLowerIR/SpecConstants.cpp#L593
However, the problem was more general and could be attributed to struct
hierarchies. `getElemDefaultValue` was essentially assuming the return
type for the spec constant instruction and the type used to initialize
the default value for the spec constant had the same "structure" (modulo
some padding fields). However, in general, this assumption does not seem
to hold. The initializer will essentially have a flattened out
structure, while the return type will retain the nested structure.

Because of this `emitSpecConstantRecursiveImpl` has been updated to
consider this. First, before recursion starts, we recurse the default
value initializer in `collectDefinedElements`, collecting information on
the children elements of the structure, noting what offset they exists
at. Then, we recurse the return type in `emitSpecConstantRecursiveImpl`.
When we reach a child element, we default value by using the information
we collected earlier. There is some consideration needed for padding -
before recursing, we check to see if any child element exists at the
offset we are trying to recurse in. If there is none, then we must be at
some padding fields.