[UR] Initial lit support
This adds initial support for lit to unified-runtime, in the form
of a new target `check-unified-runtime`. This target is available
when UR is built standalone and when built as part of dpcpp.

When built standalone, `lit` and `filecheck` must be avilable in
the path. The requirements.txt is updated to install Python
packages containing `lit` and a port of `filecheck`. Lit tests
will be run as part of the main `test` target.

When built as part of dpcpp, a new cmake variable is available;
`SYCL_UR_BUILD_TESTS` which is off by default. When enabled,
`check-unified-runtime` will be available as a target and will
also run as part of `check-all`.

Currently, only some tracing tests have been ported over to
the new system as a demonstration. Once the infrastructure is
in place, the rest of the tests will be ported.

The Python version for macos was also bumped to 3.10 to enable
use of a newer filecheck version.