[SYCL] cmd mem leak fix (#17125)
When enqueueing a command and its dependencies, an exception might be
thrown. In that case, the command will have a failed EnqueueStatus and
not stored in graph_builder DAG. But we need to make sure that
dependencies also check so they are not stored there. Otherwise there
will be leaks. During the clean up, we don't want to reenqueue it if we
know it has failed before.

This is broken out from https://github.com/intel/llvm/pull/16618