[SYCL] Ignore unknown users of virtual functions instead of crashing
The SYCLVirtualFunctionsAnalysis pass was crashing (with llvm_unreachable)
when it encountered a non-constant non-global use of an indirectly
callable function.
Instead of that just ignore any user we don't know how to handle and
log a debug message. The LLVM `User` hierarchy is mostly closed, so
in practice the ignored users are mostly going to be instructions.