[SYCL] Remove redundant check if Kernel is on host (#17376)
This is shortcut path for Kernel CG only. 
It handles only device kernels that means that there is no necessity to
check if event is a host event.
I also do not see necessity to check if Handle != nullptr since there is
a separated path for discarded event case. That means that in path for
kernel + no discarding events this check is not necessary at all.

Signed-off-by: Tikhomirova, Kseniya <kseniya.tikhomirova@intel.com>