[BENCHMARK] fix default timeout parameter (#17412)
Previously, the default value for `timeout` was bound at function
definition time using `options.timeout`, which become outdated if
`options.timeout` changes later. This commit changes the default value
to `None` and assigns the current `options.timeout` inside the function,
ensuring that the latest value is used when no explicit timeout is
provided.

Signed-off-by: Łukasz Plewa <lukasz.plewa@intel.com>