[SYCL][NewOffload][NFC] Add SYCLPostLink library component.
This change adds a component that should contain all post-link processing (e.g., ESIMD,
specialization constants). It is not appropriate to put all this functionality
into SYCLLowerIR because there is a dependency in LLVMPasses on SYCLLowerIR.
Post-link processing requires capabilities such as linking and pipeline
building, which being added to SYCLLowerIR will create a dependancy
cycle.

After this change, the dependency diagram will look like the following:
SYCLLowerIR -> Passes -> clang
                -------> opt
                |
                V
                SYCLPostLink -> clang-linker-wrapper
                |-------------> sycl-post-link
                --------------> sycl-jit.