[CI] Add GPU reset for BMG (#17381)
I talked to the KMD team and they recommended doing it this way for the
`xe` module.

Maybe this will improve stability.

I hardcoded the PCI ID but we only have one Linux BMG runner so it
should be fine for now.

Confirmed this works by checking the dmesg log of the runner.

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>