[SYCL][UR][CUDA] Fix CMake CUPTI config (#17457)
Fix-up for #17315 which used incorrect CMake variables for CUPTI:
* `CUDAToolkit_cupti_INCLUDE_DIR` should be
`CUDAToolkit_CUPTI_INCLUDE_DIR`
* `CUDAToolkit_cupti_LIBRARY` should be `CUDA_cupti_LIBRARY`

In addition, we discovered that FindCUDAToolkit:
* does not set the INCLUDE_DIR variable until CMake 3.24 (while we
support back to 3.20)
* fails to find the library in CUDA toolkits installed from runfile
(https://gitlab.kitware.com/cmake/cmake/-/issues/26770)

Fix the variable names and retry the searches with extra paths if the
header/library are not found. The REQUIRED option is enabled in the
retry, which ensures that if CUPTI is not found then the build fails at
CMake configuration instead of later during compilation or linking.