[Sycl][Graph] Reimplement topological sort algorithm
The current implementation of the topological sort
algorithm relies on a recursive function. This causes
stack overflow issues when the graphs are very large.
There is also a performance issue in the current implementation
related to the way std::find() is being used.

This commit reimplements the topological sort algorithm to fix
these issues:

- Uses an iterative approach instead of recursive.
- Removes the use of std::find() and instead relies on keeping the state
  with a helper variable.
- Reimplements the cycle checking algorithm to use the same topological
  sort function as the one used during scheduling.
- Fixes a bug with make_edge() not removing the dest node from the list of
  root nodes.