[SYCL][DOC][bindless][cuda] `gather_image` extension doc and impl (#17322)
This PR presents a portable `gather_image` API for the bindless images
extension, with an initial cuda implementation for 2D images only.
Cubemap support is also possible but not yet added at this stage, since
- it is unclear whether this can map to the spirv cube case.
- l0 doesn't yet even support cubemaps anyway

For 2D images the mapping to `__spirv_SampledImageGather` is
straightforward and unambiguous. For complete details see the extension
documentation in this PR.

---------

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>
Co-authored-by: Sean Stirling <sean.a.stirling@gmail.com>
Co-authored-by: Peter Žužek <peterzuzek@gmail.com>
Co-authored-by: Steffen Larsen <steffen.larsen@intel.com>