[SYCL] Lift restriction that prevents us from using free function kernel queries in device compilation mode (#17398)
Free function kernel queries can, at the moment, only be used in host
compilation mode requiring us to add `#ifndef __SYCL_DEVICE_ONLY__ `
directives to get code that uses them to compile. This PR removes this
restriction and the instances of the workaround used in our tests.

My understanding is the following:

The current implementation employs SFINAE using the predicate
`is_kernel_v<Func>` to delete at compile time queries that are
instantiated with a `Func` that is not a free function kernel. This is
certainly a reasonable interpretation of the spec which says :
`Constraints: Available only if is_kernel_v<Func> is true.` for each
query.

However, this causes a problem. During the device front-end compiler
phase of our toolchain, information is gathered regarding what the free
function kernels are as defined in the code provided by the user. This
means during this phase itself, `Func` will be treated as an invalid
free function kernel name since this information becomes known only
after the front-end has finished executing. And because SFINAE interacts
with the front-end mechanisms, we get a hard error in this phase of
compilation about a missing function.

One simple fix, proposed in this PR, is to simply define the trait
`is_kernel_v<Func>` to be constantly true in device compilation mode
regardless of `Func`.
This will ensure device front-end succeeds since a declaration of the
query exists and it still interprets the spec reasonably because if
`Func` does not represent a kernel, then during host compilation the
declaration will not be available because the `is_kernel` trait fails
which will print a nice message to the user saying what went wrong.