[Driver][SYCL][FPGA] Remove FPGA related offloading behaviors
Now that we have removed the option support, the next step is to clean
out the behaviors from the compiler driver.  This removes all related
support for FPGA targets.  This will not impact the existing diagnostics
that are being emitted when using the FPGA options that were recently
removed.

Remaining offloading support for AOT is for CPU and GPU targets.  As the
FPGA offloading support is only existing for the old offloading model,
this does not impact behaviors involving the new offloading model.
There are also some remaining extensions being applied during the
sycl-post-link step; this change does not touch those.