[UR] Add UR_USE_DEBUG_POSTFIX option
Instead of implicitly adding the `d` suffix to library names make this
optional. `DEBUG_POSTFIX` appears to be broken on Windows, the place
it's most useful, as `urinfo.exe` still attempts to load `ur_loader.dll`
when it should be looking for `ur_loaderd.dll`. Switching this over to
`OUTPUT_NAME ${name}d` fixes this.