[SYCL][NFC] Refactor FunctionCategorizer to simple functions.
During a code review of the Module Splitting upstreaming patch, it was recognized that the FunctionCategorizer is overengineered.
This change removes the FunctionCategorizer and replaces it with plain functions.
The rules "SimpleFlagAttribute," "SimpleFlagMetadata," and "SortedIntegersListMetadata" haven't been used and are therefore abandoned.

The aim of this change is to simplify the future upstreaming of Module Splitting.