[UR][CUDA][HIP] Fix large kernel arg buffer overflow
CUDA and HIP have a fixed maximum size for kernel arguments, this patch
ensures that passing arguments large than this maximum size is handled
properly by the adapters.

This code really needs a refactoring to avoid duplication and allow
larger argument sizes (supported by newer CUDA versions), but this will
be in follow up work.

This patch is a quick fix to avoid potential buffer overflows.