[SYCL][CUDA] Add a CUDA compatibilty mode (#12757)
This patch enables CUDA mode at the same time as the SYCL mode. This
allows SYCL code to interact with CUDA code more closely:
- A user can call a CUDA device function from a SYCL device one (follow
up of https://github.com/intel/llvm/pull/7352)
- The PR fixes overload resolution as the resolution ranking was
ambiguous in some cases
- The PR fixes error reporting, some cuda specific delayed diags weren't
reported (filtered out)
- Defines `__CUDA_ARCH__`, enabling functions to assume NVPTX is the
target

To enable the mode the user adds -fsycl-cuda-compat to the command line.
By default this mode is set to off. The flag is only used for the NVPTX
backend.

The intent is to help to transition from CUDA to SYCL. Using this mode
enable a SYCL application to reuse CUDA functionalities, especially fast
paths that are guarded by `__CUDA_ARCH__`.

---------

Signed-off-by: Victor Lomuller <victor@codeplay.com>
Co-authored-by: Tom Honermann <tom@honermann.net>