[UR] Add debug subbuild on Windows
Using ExternalProject this commit adds a subbuild of UR in compiled in
debug mode and with the `UR_USE_DEBUG_POSTFIX` option enabled. The
resulting libraries are then copied to the `<build>/bin`/`<build>/lib`
directories so they can be used in testing (not yet implemented).
Additionally, they are also included in the `deploy-sycl-toolchain`
install target alongside the normally named runtime libraries.