[SYCL][E2E] Add `build-mode` and `run-mode` to list of build-specific features (#17363)
We throw an error if a non build-specific feature is reported as true
when in build-only (we expect this to never be the case), so when we do
`UNSUPPORTED: build-only` this would make the test unresolved.

Also if we don't have `run-mode` marked as a build specific feature when
in build-only mode a test marked as `REQUIRES: run-mode` would run,
since that feature would be considered "unknown" rather than false.