[Driver] Allow for usage of old triple behavior in bundler (#17557)
Recent upstreamed changes force the triple usage with the bundler to use
a 4-field triple value.
The Intel compiler does not take well to these changes. We are not in a
position to move our triple usage when bundling and unbundling (not in a
breaking change period).

Here, we are updating bundler behavior to allow for the older behavior
of accepting triple formats that are 'incomplete'. We can move to using
the forced triple value usage when we are ready or if we even need to
when we move to the new offloading model.