[SYCL][Doc] Update GetStartedGuide pre-requisiste
This patch reformats the pre-requisites in a table (doesn't look great
in the markdown, but looks good when generated).

It also updates the required versions for CMake and Python to link to
the versions required by LLVM. For CMake we were already documenting the
same version as LLVM. For Python we didn't document any minimum version,
but we extensively use LLVM scripts like `lit`, so we really should be
aligned with it.

There's a few requirements for which we don't have minimum versions, but
we can update that if we ever run into issues with them.