[UR] Replace calls to UR in native handle functions to proper OpenCL functions (#17016)
At various points, OpenCL native handles need to be retained to ensure
SYCL semantics. Previously, this relied on the fact that UR handles were
typecast CL handles and shared the same reference count.

However, the SYCL RT shouldn't assume this, so instead we call the
appropriate (dynamically looked-up) CL functions on the native handles
instead.

This is in preperation for
https://github.com/oneapi-src/unified-runtime/pull/1176 .

This change should also have no observable effect for SYCL code; there
is no change in lifetime semantics.