Enable cmake argument passthrough in the SYCL configure wrapper (#17587)
Hundreds of `--cmake-opt=-DFOO=BAR` on the command line make it hard to
see the wood for the trees, so introduce an alternative which is to
simply pass unknown argument flags straight through to cmake. This
should make using the wrapper a little more pleasant and readable.

CHERRY-PICK OF: c62abaeee9e6cb48d04fb623075705710089bd3f

Patch-by: Luke Drummond <luke.drummond@codeplay.com>