[SYCL][Doc] Add kernel_function lambda wrapper
We previously removed the ability to pass properties directly to functions like
single_task, parallel_for, etc, in favor of requiring them to be attached to
the kernel.

By introducing a new kernel_function wrapper that attaches properties to a
kernel function, we remove the need for (most) developers to write such
wrappers themselves.

Signed-off-by: John Pennycook <john.pennycook@intel.com>