[CI][BENCHMARK] Merge new metadata implementation + tweaks to benchmarks (#17617)
This PR partially merges changes introduced in
https://github.com/intel/llvm/pull/17229. Specifically, it merges
changes related to:
- Introducing metadata in the result
- Tweaks to the benchmarks itself: benchmarks ran, parameters used to
run, what is enabled/disabled, etc
- Enables/Disables benchmarks we are interested/no longer interested in
  - Disables non-working benchmarks
  - Bumps commit hashes used to fetch each benchmark
- New command line arguments, such as `--build-jobs` and `--redownload`
(to re-fetch the benchmarks)

**Note:** I am relying on this PR having its commits squashed during
merge (which should be the default behavior for intel/llvm)

I did not try to clean up the commit history. Most of these changes are
by Piotr anyway, I wanted to make sure he was properly credited for the
changes.

Updates for current splitting effort of #17229:
https://github.com/intel/llvm/pull/17545#issuecomment-2741580146

---------

Co-authored-by: Piotr Balcer <piotr.balcer@intel.com>
Co-authored-by: Łukasz Stolarczuk <lukasz.stolarczuk@intel.com>