[SYCL][NativeCPU][NFC] Prevent generating unnamed functions
When we rename wrapper functions to take the name of the original, that
leaves in place a call to the original meaning the original cannot be
removed. When we take its name, the original's name would be cleared.
Append ".orig" instead to make it easier to keep track of which is
which.