[CI][Benchmark] Merge new static html dashboard implementation (#17659)
In continuation of the effort outlined in
https://github.com/intel/llvm/pull/17545#issuecomment-2741580146, this
PR merges further changes introduced in
https://github.com/intel/llvm/pull/17229. Specifically, it merges
@pbalcer's new HTML benchmarking results dashboard implementation, which
now loads html data from remote sources in order to facilitate bringing
up the benchmarking dashboard in intel.github.io in the future.

**Note:** I am relying on this PR having its commits squashed during
merge (which should be the default behavior for intel/llvm)

I did not try to clean up the commit history. Most of these changes are
by Piotr anyway, I wanted to make sure he was properly credited for the
changes.

---------

Co-authored-by: Piotr Balcer <piotr.balcer@intel.com>
Co-authored-by: Łukasz Stolarczuk <lukasz.stolarczuk@intel.com>