[AsyncAlloc][CUDA] Don't set pool maxSize if CUDA version < 12.2
CUDA did not introduce the maxSize property to CUmemPoolProps until
version 12.2. In the case where the user has set a maxSize > 0, and
their CUDA version is < 12.2, return an error.

This commit fixes the compilation error associated with setting this
property in older CUDA versions.

Also fix incorrect error reporting when creating a memory pool.