[UR][CUDA] Use different throttle reasons API based on CUDA version (#17719)
Our pre-commit CI uses CUDA 12.6 but nightly uses CUDA 12.1, it turns
out nvml which is part of CUDA 12.6 has
nvmlDeviceGetCurrentClocksEventReasons API, but nvml which is part of
CUDA 12.1 doesn't have it, but only supports older deprecated
nvmlDeviceGetCurrentClocksThrottleReasons.

NVML doesn't provide a version macro to check the support for that API,
so use new API nvmlDeviceGetCurrentClocksEventReasons based on cuda
version.